batch処理を追加 method update_read_count!に変更
