+++ spec/models/love_spec.rb
@@ -43,7 +43,7 @@ RSpec.describe Love, type: :model do
       end
 
       it '前日の日付分が追加される' do
-        love_with_redis_count.update_read_count(DateTime.yesterday)
+        love_with_redis_count.update_read_count!(DateTime.yesterday)
         expect(love_with_redis_count.read_count).to eq 1
         expect(Redis.current.keys).to be_empty
       end
@@ -51,7 +51,7 @@ RSpec.describe Love, type: :model do
 
     context 'love not exists read_count' do
       it '値が追加されない' do
-        love.update_read_count(DateTime.yesterday)
+        love.update_read_count!(DateTime.yesterday)
         expect(love.read_count).to eq 0
       end
     end
